rename dashboard navbar to dashboard actions, simplify statement if no funnels then load stages, move inProgres status to const
