+++ client/src/modules/layouts/Home.js
@@ -2,14 +2,14 @@ import React, { Component } from "react";
 
 import Navbar from "./Navbar/Navbar";
 import Dashboard from "../lead/Dashboard/Dashboard";
-import DashboardNavbar from "../lead/DashboardNavbar/DashboardNavbar";
+import DashboardActions from "../lead/DashboardActions/DashboardActions";
 
 class Home extends Component {
   render() {
     return (
       <div>
         <Navbar />
-        <DashboardNavbar />
+        <DashboardActions />
         <Dashboard />
       </div>
     );
